Add handling of profiles with ranks not on leaderboard
Currently the rank recalculator does not function properly if a
profile's pp is below the lowest found in the ranking data. While there
is no way to accurately recalculate rank, the rank will default to the
lowest rank found on the leaderboard, which can be inaccurate.

Hence, make it such that if new rank is below what can be found on the
leaderboard, new rank is unknown.